Fix: 공부방 종료 모달 작성 inprogress
